pyproject.toml
[tool.poetry]
name = "odoo-orm"
version = "0.1.0"
description = "A kind of Python ORM for Odoo XML-RPC API inspired by Django ORM."
authors = ["mistiru <dev@mistiru.me>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mistiru/odoo-orm"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
